+++ app/src/main/kotlin/com/maltaisn/notes/ui/edit/adapter/DragTouchHelperCallback.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Nicolas Maltais
+ * Copyright 2021 Nicolas Maltais
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -77,7 +77,7 @@ class DragTouchHelperCallback(
         viewHolder: RecyclerView.ViewHolder,
         target: RecyclerView.ViewHolder
     ): Boolean {
-        onMove(viewHolder.adapterPosition, target.adapterPosition)
+        onMove(viewHolder.bindingAdapterPosition, target.bindingAdapterPosition)
         return true
     }